feat(frontend): add settings subroutes and sidebar
