+++ frontend/src/views/settings/AlarmsSettings.vue
@@ -1,0 +1,9 @@
+<template>
+  <v-card class="pa-4">
+    <h2>Alarms Settings</h2>
+    <p>Configure alarms here.</p>
+  </v-card>
+</template>
+
+<script setup>
+</script>
