+++ frontend/src/views/settings/GatewaySettings.vue
@@ -1,0 +1,9 @@
+<template>
+  <v-card class="pa-4">
+    <h2>Gateway Settings</h2>
+    <p>Gateway configuration options.</p>
+  </v-card>
+</template>
+
+<script setup>
+</script>
