+++ frontend/src/views/settings/GeneralSettings.vue
@@ -1,0 +1,9 @@
+<template>
+  <v-card class="pa-4">
+    <h2>General Settings</h2>
+    <p>Configure general parameters here.</p>
+  </v-card>
+</template>
+
+<script setup>
+</script>
