+++ frontend/src/views/settings/GroupsSettings.vue
@@ -1,0 +1,9 @@
+<template>
+  <v-card class="pa-4">
+    <h2>Groups Settings</h2>
+    <p>Manage your groups here.</p>
+  </v-card>
+</template>
+
+<script setup>
+</script>
