+++ frontend/src/views/settings/NodesSettings.vue
@@ -1,0 +1,9 @@
+<template>
+  <v-card class="pa-4">
+    <h2>Nodes Settings</h2>
+    <p>Nodes configuration will appear here.</p>
+  </v-card>
+</template>
+
+<script setup>
+</script>
