+++ frontend/src/views/settings/ProfilesSettings.vue
@@ -1,0 +1,9 @@
+<template>
+  <v-card class="pa-4">
+    <h2>Profiles Settings</h2>
+    <p>User profiles configuration.</p>
+  </v-card>
+</template>
+
+<script setup>
+</script>
