+++ frontend/src/views/settings/SystemSettings.vue
@@ -1,0 +1,9 @@
+<template>
+  <v-card class="pa-4">
+    <h2>System Settings</h2>
+    <p>System-wide configuration goes here.</p>
+  </v-card>
+</template>
+
+<script setup>
+</script>
